+++ src/components/applications/ApplicationRowItem.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import { truncateHash } from '../../utils/displayFunctions';
-import { Application } from '../../api/dataSource/NodeDataSource';
 import { ClipboardDocumentIcon } from '@heroicons/react/24/solid';
+import { Application } from '../../pages/Applications';
 
 interface ApplicationRowItemProps {
   $hasBorders: boolean;